canvas: MetaballCanvas + metaball action
WebGL metaball simulation ported to Svelte action (actions/metaball.ts).
Uses update() callback pattern (not $effect) so target/trackCursor/color
props update live without tearing down the GL context.

gl narrowed to non-null alias after early-return guard for clean closures.
